refactor: import 컨벤션 준수 및 useClickOutside 훅 활용 JUL-46
