+++ app/my-profile/edit/page.tsx
@@ -1,5 +1,5 @@
-import Link from "next/link";
 import Image from "next/image";
+import Link from "next/link";
 import { EditProfile } from "components/employee";
 import styles from "./page.module.scss";
