+++ app/my-profile/page.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { CommonDetail, CommonLayout, StatusChip } from "components/common";
-import RegisteredMyProfile from "components/employee/RegisteredMyProfile/RegisteredMyProfile";
 import { EmployeeTable } from "components/notice";
+import RegisteredMyProfile from "components/employee/RegisteredMyProfile/RegisteredMyProfile";
 import { DetailType } from "types/enums/detailPage.enum";
 import { ApplyStatus } from "types/enums/apply.enum";
 import styles from "./page.module.scss";
