+++ components/auth/AuthForm/AuthForm.tsx
@@ -3,8 +3,8 @@
 import {
   useEffect, useMemo, useState,
 } from "react";
-import { usePathname, useSearchParams } from "next/navigation";
 import Link from "next/link";
+import { usePathname, useSearchParams } from "next/navigation";
 import classNames from "classnames/bind";
 import SigninForm from "./SigninForm";
 import SignupForm from "./SignupForm";
