+++ components/common/GlobalFooter/GlobalFooter.tsx
@@ -1,5 +1,5 @@
-import Link from "next/link";
 import Image from "next/image";
+import Link from "next/link";
 import styles from "./GlobalFooter.module.scss";
 
 const GlobalFooter = () => {
