+++ components/common/PostCard/PostCard.tsx
@@ -1,8 +1,8 @@
-import Link from "next/link";
 import Image from "next/image";
+import Link from "next/link";
 import classNames from "classnames/bind";
-import formatTimeRange from "utils/formatTimeRange";
 import calculatePercentage from "utils/calculatePercentage";
+import formatTimeRange from "utils/formatTimeRange";
 import getBgColorClass from "utils/getBgColorClass";
 import CustomArrow from "./CustomArrow/CustomArrow";
 import styles from "./PostCard.module.scss";
