+++ components/common/SearchBar/SearchBar.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { useState } from "react";
-import { useRouter } from "next/navigation";
 import Image from "next/image";
+import { useRouter } from "next/navigation";
 import styles from "./SearchBar.module.scss";
 
 interface SearchBarProps {
