+++ components/notice/EmployerTable/EmployerTable.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import Table from "components/common/Table/Table";
-import { EMPLOYER_COLUMNS } from "constants/notice";
 import { IEmployerNotices } from "types/notice/tables";
+import { EMPLOYER_COLUMNS } from "constants/notice";
 
 interface EmployerTableProps {
   applicationList: IEmployerNotices[];
