+++ components/notice/FilterButton/FilterButton.tsx
@@ -1,29 +1,18 @@
 "use client";
 
-import { useState, useEffect, useRef } from "react";
+import { useState, useRef } from "react";
 import { Filter } from "components/common";
+import useClickOutside from "hooks/useClickOutside";
 import styles from "./FilterButton.module.scss";
 
 const FilterButton = () => {
   const [showPopover, setShowPopover] = useState(false);
   const containerRef = useRef<HTMLDivElement>(null);
+  useClickOutside(containerRef, () => { return setShowPopover(false); });
 
   const toggleShowPopup = () => {
     setShowPopover((prev) => { return !prev; });
   };
-
-  const handleOutsideClick = (e: MouseEvent) => {
-    if (containerRef.current && !containerRef.current.contains(e.target as Node)) {
-      setShowPopover(false);
-    }
-  };
-
-  useEffect(() => {
-    document.addEventListener("mousedown", handleOutsideClick);
-    return () => {
-      document.removeEventListener("mousedown", handleOutsideClick);
-    };
-  }, []);
 
   return (
     <div
